requirements.txt
pyspark==3.5.0
pandas==2.2.2
matplotlib==3.8.4
seaborn==0.13.2
pyarrow==15.0.2
notebook==7.1.2
jupyterlab==4.1.5
pyyaml==6.0.1
scipy==1.13.0
statsmodels==0.14.0

